add product api bug is resolved

+++ api/src/main/java/com/community/api/dto/AddProductDto.java
@@ -97,7 +97,7 @@ public class AddProductDto {
     Long formComplexity;
 
     @JsonProperty("qualification_id")
-    Long qualification;
+    Integer qualification;
     @JsonProperty("stream_id")
     Long stream;
     @JsonProperty("subject_id")
